Cleanup import ordering
