+++ internal/util/testing.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/NETWAYS/support-collector/internal/obfuscate"
+
 	"github.com/stretchr/testify/assert"
 )
