+++ internal/util/testing_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/NETWAYS/support-collector/internal/obfuscate"
+
 	"github.com/stretchr/testify/assert"
 )
