+++ modules/base/collector.go
@@ -6,6 +6,7 @@ import (
 	"path/filepath"
 
 	"github.com/NETWAYS/support-collector/internal/collection"
+
 	"gopkg.in/yaml.v3"
 )
