+++ modules/base/kernel_test.go
@@ -1,8 +1,9 @@
 package base
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestGetKernelInfo(t *testing.T) {
