+++ modules/graylog/collector.go
@@ -1,11 +1,11 @@
 package graylog
 
 import (
-	"github.com/NETWAYS/support-collector/internal/util"
 	"path/filepath"
 
 	"github.com/NETWAYS/support-collector/internal/collection"
 	"github.com/NETWAYS/support-collector/internal/obfuscate"
+	"github.com/NETWAYS/support-collector/internal/util"
 )
 
 const ModuleName = "graylog"
