+++ modules/prometheus/collector_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/NETWAYS/support-collector/internal/collection"
 	"github.com/NETWAYS/support-collector/internal/obfuscate"
 	"github.com/NETWAYS/support-collector/internal/util"
+
 	"github.com/stretchr/testify/assert"
 )
